split miner to sector and miner
